python-service/requirements.txt
fastapi
uvicorn[standard]
aiokafka
motor
pydantic
PyJWT
python-dotenv
cryptography
